refactor(watch): extract useChannelFetch and de-duplicate demo/real fetch paths
